+++ CHANGES.rst
@@ -24,6 +24,7 @@ Development
 - Fix network status flag
 - Improve cross-compilation tooling re. Pycom vs. pyboard-D
 - Display HTTP egress payload in log output on INFO level
+- Add option to use external WiFi antenna
 
 
 2019-08-19 0.6.0